+++ crates/config/tests/utils.rs
@@ -152,6 +152,9 @@ pub fn load_project_config_in_format(format: &str) {
                     exclude: vec![Id::raw("build")],
                     include: Some(vec![Id::raw("test")]),
                     rename: FxHashMap::from_iter([(Id::raw("old"), Id::raw("new"))])
+                },
+                merge_strategies: ProjectWorkspaceMergeStrategiesConfig {
+                    file_groups: Some(MergeStrategy::Replace)
                 }
             },
             ..Default::default()
@@ -293,11 +296,11 @@ pub fn load_tasks_config_in_format(format: &str) {
                 interactive: Some(false),
                 internal: Some(true),
                 merge: None,
-                merge_args: Some(TaskMergeStrategy::Append),
+                merge_args: Some(MergeStrategy::Append),
                 merge_checks: None,
-                merge_deps: Some(TaskMergeStrategy::Prepend),
-                merge_env: Some(TaskMergeStrategy::Replace),
-                merge_inputs: Some(TaskMergeStrategy::Preserve),
+                merge_deps: Some(MergeStrategy::Prepend),
+                merge_env: Some(MergeStrategy::Replace),
+                merge_inputs: Some(MergeStrategy::Preserve),
                 merge_outputs: None,
                 merge_tags: None,
                 merge_toolchains: None,